backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-environ==0.11.2
graphql-core==3.2.3
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
strawberry-graphql==0.227.3
strawberry-graphql-django==0.39.2
typing_extensions==4.11.0
